client server restructured, css works with react components, static pics works, pages flow works well
