+++ client/components/cards/Cardpicdetails.js
