+++ client/css/style.css
@@ -1,3 +1,12 @@
 h1 {
-  color: "red";
-  background-color: yellow; }
+  	color: red;
+  	background-color: yellow; 
+}
+
+.contactAccess {
+	color: white;
+	background-color: red;
+	cursor: pointer;
+	width: 150px;
+}
+
