+++ client/pages/contact/index.js
@@ -1,3 +1,11 @@
 export default() => {
-	return <h1> Contact page !!!</h1>
+	return (
+		<div>
+		<h1> Contact page !!!</h1>
+		<img src="/contact/qrcode.png" />
+		<img src="/contact/facebook.png" />
+		<img src="/contact/line.png" />
+		<img src="/contact/mail.jpeg" />
+		</div>
+	)
 }
