+++ client/pages/type/[typeId].js
@@ -2,7 +2,8 @@ import Router from 'next/router'
 import renderBanner from '../../services/renderBanner'
 import translator from '../../services/translator'
 
-import Cardstd from '../../components/cardStd/Cardstd' 
+import Cardstd from '../../components/cards/Cardstd' 
+import Gotocontact from '../../components/cards/cardGlobal/Gotocontact'
 
 const TypePage = ({ pics, type }) => {
 	console.log('picsmain from type: ', pics)
@@ -15,11 +16,14 @@ const TypePage = ({ pics, type }) => {
 
 	const carsList = pics.carsData.map(car => {
 		return(
-			<Cardstd
-				key={car._id}
-				carDetails={car}
-				clicked={() => goToCar(`${car._id}=${car.style}`)}
-			/>
+			<div>
+				<Cardstd
+					key={car._id}
+					carDetails={car}
+					clicked={() => goToCar(`${car._id}=${car.style}`)}
+				/>
+				<Gotocontact />
+			</div>
 		)
 	})
